refactor [#25] Sign up 및 비밀번호 찾기 화면들을 만들 때 사용하는 공통 코드를 추출
